moved EmployeeComponent.jsx to Components folder
