+++ src/Components/EmployeeComponent/EmployeeComponent.css
@@ -1,0 +1,3 @@
+.table-element{
+    padding-top: 10px;
+}
